+++ matomo/plugins/QueuedTracking/Commands/PrintQueuedRequests.php
@@ -9,12 +9,9 @@
 
 namespace Piwik\Plugins\QueuedTracking\Commands;
 
-use Piwik\Plugin;
 use Piwik\Plugin\ConsoleCommand;
 use Piwik\Plugins\QueuedTracking\Queue;
-use Piwik\Plugins\QueuedTracking\Queue\Processor;
 use Piwik\Plugins\QueuedTracking\SystemCheck;
-use Piwik\Tracker;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -57,5 +54,7 @@ class PrintQueuedRequests extends ConsoleCommand
             $output->writeln(sprintf('<info>These were the requests of queue %s. Use <comment>--queue-id=%s</comment> to print only information for this queue.</info>', $thisQueueId, $thisQueueId));
 
         }
+
+        return 0;
     }
 }