+++ matomo/plugins/QueuedTracking/SystemSettings.php
@@ -8,12 +8,9 @@
  */
 namespace Piwik\Plugins\QueuedTracking;
 
-use Piwik\Cache;
-use Piwik\Config;
 use Piwik\Plugins\QueuedTracking\Settings\NumWorkers;
 use Piwik\Settings\Setting;
 use Piwik\Settings\FieldConfig;
-use Piwik\Settings\Storage\Backend;
 use Piwik\Plugins\QueuedTracking\Queue\Factory;
 use Piwik\Piwik;
 use Exception;
@@ -231,7 +228,7 @@ class SystemSettings extends \Piwik\Settings\Plugin\SystemSettings
             $field->uiControlAttributes = array('size' => 128);
             $field->inlineHelp = 'Password set on the Redis server, if any. Redis can be instructed to require a password before allowing clients to execute commands.';
             $field->validate = function ($value) {
-                if (strlen($value) > 128) {
+                if (is_string($value) && strlen($value) > 128) {
                     throw new \Exception('Max 128 characters allowed');
                 }
             };
@@ -265,7 +262,7 @@ class SystemSettings extends \Piwik\Settings\Plugin\SystemSettings
         return $this->makeSetting('queueEnabled', $default = false, FieldConfig::TYPE_BOOL, function (FieldConfig $field) use ($self) {
             $field->title = 'Queue enabled';
             $field->uiControl = FieldConfig::UI_CONTROL_CHECKBOX;
-            $field->inlineHelp = 'If enabled, all tracking requests will be written into a queue instead of the directly into the database. Requires a Redis server and phpredis PHP extension.';
+            $field->inlineHelp = 'If enabled, all tracking requests will be written into a queue instead of the directly into the database. Requires a Redis server and phpredis PHP extension if using Redis as a backend.';
             $field->validate = function ($value) use ($self) {
                 $value = (bool) $value;