+++ matomo/plugins/QueuedTracking/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+4.0.4
+- Clarify inline help for "Queue enabled" config setting
+
+4.0.3
+- Replace Redis::delete() with Redis::del() and fix a warning
+
 4.0.2
 - Support new option `--force-num-requests-process-at-once` to the process command